Refactored soil sample import
Less separate database calls. All one function to do all the appropriate inserts. This is better because creating a soil sample is one operation that leads to multiple tables in the database. None of the subordinates should ever be called individually

+++ SoilApp/SoilDB/dbo/Stored Procedures/spFieldsNutrients_Upsert.sql
@@ -1,7 +1,7 @@
 ﻿CREATE PROCEDURE [dbo].[spFieldsNutrients_Upsert]
 	@FieldId int,
 	@NutrientId int,
-	@Amount int,
+	@Amount decimal(10,5),
 	@Goal int,
 	@SampleYear int,
 	@id int=0 output
